requirements.txt
numpy==1.26.3
opencv-python==4.10.0.84
Pillow==10.1.0
torch==2.1.2
onnxruntime==1.19.2
tqdm==4.66.1